--- src/apis/supplyUsd.tsx
@@ -1,10 +1,0 @@
-import cointousd from "./cointousd";
-import {
-  aave_address,
-  aave_oracle_contract,
-  dai_address,
-  dai_contract,
-  usdc_address,
-} from "./common";
-
-export async function DaiToUsd() {}
