--- src/app/account/page.tsx
@@ -1,14 +1,0 @@
-"use client";
-
-import Background from "@/components/layout/Background";
-import Header from "@/components/layout/Header";
-import AccountMain from "@/components/account/AccountMain";
-
-export default function Account() {
-  return (
-    <Background>
-      <Header />
-      <AccountMain />
-    </Background>
-  );
-}
