+++ src/app/layout.tsx
@@ -1,13 +1,9 @@
 import { ReactNode } from "react";
-import { TmpContextProvider } from "@/components/TmpContext";
-import "@/styles/globals.css";
 
 export default function RootLayout({ children }: { children: ReactNode }) {
   return (
-    <html lang="en">
-      <body>
-        <TmpContextProvider>{children}</TmpContextProvider>
-      </body>
+    <html lang="en" className="h-full w-full">
+      <body className="h-full w-full">{children}</body>
     </html>
   );
 }
