+++ src/app/page.tsx
@@ -1,14 +1,3 @@
-"use client";
-
-import Background from "../components/layout/Background";
-import Header from "../components/layout/Header";
-import Main from "../components/main/Main";
-
-export default function Home() {
-  return (
-    <Background>
-      <Header />
-      <Main />
-    </Background>
-  );
+export default function LandingPage() {
+  return <p className="text-white">TODO: landing</p>;
 }
