--- src/app/reward/page.tsx
@@ -1,14 +1,0 @@
-"use client";
-
-import Background from "@/components/layout/Background";
-import RewardMain from "@/components/reward/RewardMain";
-import Header from "@/components/layout/Header";
-
-export default function Reward() {
-  return (
-    <Background>
-      <Header />
-      <RewardMain />
-    </Background>
-  );
-}
