+++ src/components/main/supply/SupplyTableField.tsx
@@ -4,22 +4,20 @@ interface Props {
 
 export default function SupplyTableField({ tableRow }: Props) {
   return (
-    <>
-      <tr>
-        {tableRow &&
-          tableRow.map((item: string, index: number) => (
-            <th
-              key={index}
-              className={
-                index === 5
-                  ? `font-pretendard text-[14px] font-semibold leading-[20px] text-[#559465]`
-                  : `font-pretendard text-[14px] font-semibold leading-[20px] text-[#525C52]`
-              }
-            >
-              {item}
-            </th>
-          ))}
-      </tr>
-    </>
+    <tr>
+      {tableRow &&
+        tableRow.map((item: string, index: number) => (
+          <th
+            key={index}
+            className={
+              index === 5
+                ? `font-pretendard text-[14px] font-semibold leading-[20px] text-[#559465]`
+                : `font-pretendard text-[14px] font-semibold leading-[20px] text-[#525C52]`
+            }
+          >
+            {item}
+          </th>
+        ))}
+    </tr>
   );
 }
